Feature/recipes

+++ src/main/java/com/restock/platform/planning/interfaces/rest/resources/RecipeSupplyResource.java
@@ -1,3 +1,3 @@
 package com.restock.platform.planning.interfaces.rest.resources;
 
-public record RecipeSupplyResource(Long supplyId, Double quantity) {}
+public record RecipeSupplyResource(Integer supplyId, Double quantity) {}
